Drop PromiseKit dependency from AwfulCore and take apart ForumsURLSession
The Awful app target was implicitly relying on AwfulCore's depending on PromiseKit. Now the app dependency is explicit.
