+++ AwfulCore/Package.swift
@@ -14,7 +14,6 @@ let package = Package(
         .package(path: "../AwfulScraping"),
         .package(path: "../AwfulSwift"),
         .package(path: "../Logger"),
-        .package(url: "https://github.com/mxcl/PromiseKit", .upToNextMajor(from: "6.0.0")),
         .package(url: "https://github.com/nolanw/HTMLReader", .upToNextMajor(from: "2.1.7")),
     ],
     targets: [
@@ -25,7 +24,6 @@ let package = Package(
                 "AwfulSwift",
                 "HTMLReader",
                 "Logger",
-                "PromiseKit",
             ],
             resources: [
                 .process("Localizable.strings"),
